rename contracts/sdk to helper
